chore: moved new() and load() functions out the Corporeum struct
With this change, users are not able to call Corporeum::save() or Corporeum::corpus(), before they actually create/load something.
Now the users can either create a new corporeum struct or load an existing one.
